requirements.txt
streamlit>=1.34.0
pycryptodome>=3.20.0
